Pulling out view to make the code reusable, more readable and better for future testing.
