+++ FeedbackHub/AwardsView.swift
@@ -39,9 +39,9 @@ struct AwardsView: View {
                                 .scaledToFit()
                                 .padding()
                                 .frame(width: 100, height: 100)
-                                .foregroundColor(dataController.hasEarned(award: award) ? Color(award.color) : .secondary.opacity(0.5))
+                                .foregroundColor(color(for: award))
                         }
-                        .accessibilityLabel(dataController.hasEarned(award: award) ? "Unlocked: \(award.name)" : "Locked")
+                        .accessibilityLabel(label(for: award))
                         .accessibilityHint(award.description)
                     }
                 }
@@ -52,6 +52,14 @@ struct AwardsView: View {
                 Text(selectedAward.description)
             }
         }
+    }
+    
+    func color(for award: Award) -> Color {
+        dataController.hasEarned(award: award) ? Color(award.color) : .secondary.opacity(0.5)
+    }
+    
+    func label(for award: Award) -> LocalizedStringKey {
+        dataController.hasEarned(award: award) ? "Unlocked: \(award.name)" : "Locked"
     }
 }
