+++ FeedbackHub/IssueView.swift
@@ -30,37 +30,8 @@ struct IssueView: View {
                     Text("Medium").tag(Int16(1))
                     Text("High").tag(Int16(2))
                 }
-
-                Menu {
-                    // show selected tags first
-                    ForEach(issue.issueTags) { tag in
-                        Button {
-                            issue.removeFromTags(tag)
-                        } label: {
-                            Label(tag.tagName, systemImage: "checkmark")
-                        }
-                    }
-
-                    // now show unselected tags
-                    let otherTags = dataController.missingTags(from: issue)
-
-                    if otherTags.isEmpty == false {
-                        Divider()
-
-                        Section("Add Tags") {
-                            ForEach(otherTags) { tag in
-                                Button(tag.tagName) {
-                                    issue.addToTags(tag)
-                                }
-                            }
-                        }
-                    }
-                } label: {
-                    Text(issue.issueTagsList)
-                        .multilineTextAlignment(.leading)
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                        .animation(nil, value: issue.issueTagsList)
-                }
+                
+               TagsMenuView(issue: issue)
             }
 
             Section {
@@ -79,23 +50,7 @@ struct IssueView: View {
         }
         .onSubmit(dataController.save)
         .toolbar {
-            Menu {
-                Button {
-                    UIPasteboard.general.string = issue.title
-                } label: {
-                    Label("Copy Issue Title", systemImage: "doc.on.doc")
-                }
-                
-                Button {
-                    issue.completed.toggle()
-                    dataController.save()
-                } label: {
-                    Label(issue.completed ? "Re-open Issue" : "Closed Issue", systemImage: "bubble.left.and.exclamationmark.bubble.right")
-                }
-                
-            } label: {
-                Label("Action", systemImage: "ellipsis.circle")
-            }
+            IssueViewToolbar(issue: issue)
         }
     }
 }
